fix(gui): the LRO knob row index compiles on Windows
The v3.0.0-beta.0 tag build: macOS green, Windows red on one line.

    error C2672: 'juce::jmax': no matching overloaded function found
    error C2737: 'idx': const object must be initialized
    src/gui/SynthPanel.cpp(1563)

`juce::String::operator[]` returns `juce_wchar`, and that type is not the same
everywhere: a signed 32-bit `wchar_t` on macOS and Linux, `juce::uint32` on
Windows, where `wchar_t` is UTF-16. So `k.slot[0] - 'a'` is `int` on two
platforms and `unsigned int` on the third, `juce::jmax(0, ...)` cannot deduce
one `Type` from an `int` and an `unsigned int`, and MSVC rejects the whole
declaration. Nothing about the line looks platform-specific, which is why it
passed every local build.

The cast moves in front of the subtraction rather than around the result, which
also restores what the `jmax` was for: with an unsigned difference a slot below
'a' would wrap to a huge positive instead of going negative, so the clamp was
dead on Windows and only the `idx >= kNumLroKnobs` guard below stood between
that and a wrong knob row.

+++ src/gui/SynthPanel.cpp
@@ -1560,7 +1560,15 @@ void SynthPanel::layoutLroKnobs()
         const auto knobs = lroControls.knobsOf(lroControls.parts[c].number);
         for (const auto& k : knobs)
         {
-            const int idx = (k.part - 1) * 4 + juce::jmax(0, k.slot[0] - 'a');
+            // static_cast before the subtraction, not after. `juce::String::operator[]`
+            // returns `juce_wchar`, which is a SIGNED 32-bit wchar_t on macOS and
+            // Linux and `juce::uint32` on Windows — so `k.slot[0] - 'a'` is `int`
+            // there and `unsigned int` here, `jmax(0, …)` cannot deduce one type, and
+            // MSVC refuses the whole line. Unsigned would also have made the `jmax`
+            // clamp dead: a slot below 'a' would wrap to a huge positive instead of
+            // going negative.
+            const int idx = (k.part - 1) * 4
+                          + juce::jmax(0, static_cast<int>(k.slot[0]) - 'a');
             if (idx < 0 || idx >= kNumLroKnobs || lroKnobRows[static_cast<size_t>(idx)] == nullptr)
                 continue;
             auto& row = *lroKnobRows[static_cast<size_t>(idx)];
